Rename, stupidly, lockfiles
Instead of calling them "foo.lock", which sometimes conflicts (e.g.
"Gemfile.lock" is a thing), we call them "foo..lg2", which is much
less likely to conflict.

My test repo actually has some cases of a dir legitimiately containing
files "list" and "list.lock" (I think it's a golang thing)

This has a double-dot, meaning that it won't cause issues with refs,
and it is excedingly unlikely to be an actual filename.  A smart fix
would continue to use ".lock" for refs (since it's reserved), but
this is just a temp fix until upstream fixes things.